refactor(groups): extraer ShareGroupDialog y ajustar flujo de abandonar grupo
