+++ src/features/groups/hooks/useLeaveGroup.ts
@@ -4,24 +4,31 @@ import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { useTranslations } from 'next-intl';
 import { api } from '@lib/api/instance';
+import useSnackbar from '@shared/hooks/useSnackbar';
 import type { LeaveGroupResponse } from '../types';
 
 interface UseLeaveGroupReturn {
   loading: boolean;
   error: string | null;
-  leaveGroup: (groupId: string, dissolve?: boolean) => Promise<LeaveGroupResponse | null>;
+  leaveGroup: (
+    groupId: string,
+    dissolve?: boolean,
+    assigningNewAdmin?: boolean,
+  ) => Promise<LeaveGroupResponse | null>;
   resetError: () => void;
 }
 
 export const useLeaveGroup = (): UseLeaveGroupReturn => {
   const router = useRouter();
   const t = useTranslations('groups');
+  const snackbar = useSnackbar();
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
 
   const leaveGroup = async (
     groupId: string,
     dissolve?: boolean,
+    assigningNewAdmin?: boolean,
   ): Promise<LeaveGroupResponse | null> => {
     setLoading(true);
     setError(null);
@@ -31,6 +38,15 @@ export const useLeaveGroup = (): UseLeaveGroupReturn => {
 
       if (isMock) {
         await new Promise((resolve) => setTimeout(resolve, 600));
+        snackbar.success(
+          t(
+            dissolve
+              ? 'dissolveSuccess'
+              : assigningNewAdmin
+                ? 'leaveSuccessNewAdmin'
+                : 'leaveSuccess',
+          ),
+        );
         router.push('/home');
         return { success: true, dissolved: dissolve ?? false };
       }
@@ -45,6 +61,15 @@ export const useLeaveGroup = (): UseLeaveGroupReturn => {
         return null;
       }
 
+      snackbar.success(
+        t(
+          dissolve
+            ? 'dissolveSuccess'
+            : assigningNewAdmin
+              ? 'leaveSuccessNewAdmin'
+              : 'leaveSuccess',
+        ),
+      );
       router.push('/home');
       return response.data ?? { success: true };
     } catch {
